+++ tests/testthat/test_center_scale_norm.R
@@ -141,3 +141,26 @@ test_that('correct means and std devs for step_norm', {
 
   expect_equal(tidy(standardized_trained, 1), norm_tibble_tr)
 })
+
+test_that('scale by factor of 1 or 2 for step_normalize', {
+  standardized <- rec %>%
+    step_normalize(carbon, hydrogen, oxygen, nitrogen, sulfur, id = "scale", factor = 2)
+
+  standardized_trained <- prep(standardized, training = biomass)
+
+  scal_tibble_tr <-
+    tibble(terms = c("carbon", "hydrogen", "oxygen", "nitrogen", "sulfur"),
+           value = sds*2,
+           id = standardized$steps[[1]]$id)
+
+  expect_equal(tidy(standardized_trained, 1), scal_tibble_tr)
+
+  expect_equal(standardized_trained$steps[[1]]$sds, 2*sds)
+
+  expect_warning(
+    not_recommended_standardized_input <- rec %>%
+      step_normalize(carbon, id = "scale", factor = 3) %>%
+      prep(training = biomass)
+  )
+
+})